LayoutTests/media/track/track-webvtt-tc018-align-text-line-position-expected.txt

Tests cue alignment, line and text position from settings.


*** Testing text track 0
EXPECTED (cues.length == '5') OK
EXPECTED (cues[0].alignment == 'start') OK
EXPECTED (cues[0].textPosition == '10') OK
EXPECTED (cues[0].linePosition == '0') OK
EXPECTED (cues[0].snapToLines == 'false') OK
EXPECTED (cues[1].alignment == 'start') OK
EXPECTED (cues[1].textPosition == '50') OK
EXPECTED (cues[1].linePosition == '0') OK
EXPECTED (cues[1].snapToLines == 'true') OK
EXPECTED (cues[2].alignment == 'middle') OK
EXPECTED (cues[2].textPosition == '80') OK
EXPECTED (cues[2].linePosition == '80') OK
EXPECTED (cues[2].snapToLines == 'false') OK
EXPECTED (cues[3].alignment == 'end') OK
EXPECTED (cues[3].textPosition == '30') OK
EXPECTED (cues[3].linePosition == '5') OK
EXPECTED (cues[3].snapToLines == 'true') OK
EXPECTED (cues[4].alignment == 'middle') OK
EXPECTED (cues[4].textPosition == '60') OK
EXPECTED (cues[4].linePosition == '-3') OK
EXPECTED (cues[4].snapToLines == 'true') OK

*** Testing text track 1
EXPECTED (cues.length == '3') OK
EXPECTED (cues[0].alignment == 'middle') OK
EXPECTED (cues[0].textPosition == '50') OK
EXPECTED (cues[0].linePosition == '-1') OK
EXPECTED (cues[0].snapToLines == 'true') OK
EXPECTED (cues[1].alignment == 'end') OK
EXPECTED (cues[1].textPosition == '0') OK
EXPECTED (cues[1].linePosition == '-1') OK
EXPECTED (cues[1].snapToLines == 'true') OK
EXPECTED (cues[2].alignment == 'middle') OK
EXPECTED (cues[2].textPosition == '60') OK
EXPECTED (cues[2].linePosition == '-3') OK
EXPECTED (cues[2].snapToLines == 'true') OK
END OF TEST
